refactor: `LikeService`의 기능을 `ContentService`로 옮긴다
