+++ back/src/main/java/com/baba/back/content/controller/ContentController.java
@@ -1,10 +1,9 @@
 package com.baba.back.content.controller;
 
-import com.baba.back.content.dto.AddLikeResponse;
+import com.baba.back.content.dto.LikeContentResponse;
 import com.baba.back.content.dto.CreateContentRequest;
 import com.baba.back.content.dto.CreateContentResponse;
 import com.baba.back.content.service.ContentService;
-import com.baba.back.content.service.LikeService;
 import com.baba.back.oauth.support.Login;
 import com.baba.back.swagger.BadRequestResponse;
 import com.baba.back.swagger.CreatedResponse;
@@ -30,7 +29,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class ContentController {
 
     private final ContentService contentService;
-    private final LikeService likeService;
 
     @Operation(summary = "컨텐츠 생성 요청")
     @CreatedResponse
@@ -53,8 +51,8 @@ public class ContentController {
     @NotFoundResponse
     @IntervalServerErrorResponse
     @PostMapping("/album/{babyId}/{contentId}/like")
-    public ResponseEntity<AddLikeResponse> addLike(@Login String memberId, @PathVariable("babyId") String babyId,
-                                                   @PathVariable("contentId") Long contentId) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(likeService.addLike(memberId, babyId, contentId));
+    public ResponseEntity<LikeContentResponse> addLike(@Login String memberId, @PathVariable("babyId") String babyId,
+                                                       @PathVariable("contentId") Long contentId) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(contentService.likeContent(memberId, babyId, contentId));
     }
 }
