+++ back/src/main/java/com/baba/back/content/dto/LikeContentResponse.java
@@ -3,5 +3,5 @@ package com.baba.back.content.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record AddLikeResponse(@JsonProperty("isLiked") boolean liked) {
+public record LikeContentResponse(@JsonProperty("isLiked") boolean liked) {
 }
