+++ back/src/main/java/com/baba/back/content/repository/LikeRepository.java
@@ -1,7 +1,11 @@
 package com.baba.back.content.repository;
 
 import com.baba.back.content.domain.Like;
+import com.baba.back.content.domain.content.Content;
+import com.baba.back.oauth.domain.member.Member;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
+    Optional<Like> findByMemberAndContent(Member member, Content content);
 }
